Revert "wip example of tanstack query for responsive ux"
This reverts commit 1773312e7b64380e35b2d1dabbc376b1d5faf3d4.
