+++ app/[locale]/[org]/[project]/settings/_components/AllowAnonymousRecruitmentSwitch.tsx
@@ -1,38 +1,21 @@
-import { getAnonymousRecruitmentStatus } from '~/server/actions/projects';
+'use server';
+import { getAnonymousRecruitmentStatus } from '~/server/queries/projects';
 import SwitchWithOptimisticUpdate from '~/components/SwitchWithOptimisticUpdate';
 import { setAnonymousRecruitment } from '~/server/actions/projects';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
-const AnonymousRecruitmentSwitch = ({
+const AnonymousRecruitmentSwitch = async ({
   projectSlug,
 }: {
   projectSlug: string;
 }) => {
-  const { data: allowAnonymousRecruitment } = useQuery({
-    queryKey: ['allowAnonymousRecruitment', projectSlug],
-    queryFn: () => getAnonymousRecruitmentStatus(projectSlug),
-  });
-
-  const queryClient = useQueryClient();
-
-  const { mutate } = useMutation({
-    mutationFn: setAnonymousRecruitment.bind(null, projectSlug),
-    onSettled: () =>
-      queryClient.invalidateQueries({
-        queryKey: ['allowAnonymousRecruitment', projectSlug],
-      }),
-    mutationKey: ['setAnonymousRecruitment'],
-  });
-
-  if (allowAnonymousRecruitment === undefined) {
-    return null;
-  }
+  const allowAnonymousRecruitment =
+    await getAnonymousRecruitmentStatus(projectSlug);
 
   return (
     <SwitchWithOptimisticUpdate
       initialValue={allowAnonymousRecruitment}
       name="allowAnonymousRecruitment"
-      mutation={mutate}
+      action={setAnonymousRecruitment.bind(null, projectSlug)}
     />
   );
 };
