+++ package.json
@@ -26,7 +26,6 @@
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.7",
-    "@tanstack/react-query": "^5.44.0",
     "@typescript-eslint/eslint-plugin": "^7.12.0",
     "@typescript-eslint/parser": "^7.12.0",
     "clsx": "^2.1.1",
@@ -62,7 +61,6 @@
     "@storybook/nextjs": "^8.1.6",
     "@storybook/react": "^8.1.6",
     "@storybook/test": "^8.1.6",
-    "@tanstack/eslint-plugin-query": "^5.43.1",
     "@total-typescript/ts-reset": "^0.5.1",
     "@types/node": "^20",
     "@types/react": "^18",
