+++ server/actions/projects.ts
@@ -40,9 +40,6 @@ export async function setAnonymousRecruitment(
   input: boolean,
 ) {
   try {
-    // Simulate a delay to test optimistic UI
-    await new Promise((resolve) => setTimeout(resolve, 2000));
-
     await db
       .update(projects)
       .set({ allowAnonymousRecruitment: input })
@@ -56,15 +53,4 @@ export async function setAnonymousRecruitment(
     console.error('Unable to update project anonymous recruitment status', e);
     return false;
   }
-}
-
-export async function getAnonymousRecruitmentStatus(projectSlug: string) {
-  const project = await db.query.projects.findFirst({
-    where: eq(projects.slug, projectSlug),
-    columns: {
-      allowAnonymousRecruitment: true,
-    },
-  });
-
-  return !!project?.allowAnonymousRecruitment;
 }
